JDBI initialization move to common
