+++ backend/src/main/java/ai/verta/modeldb/config/TestConfig.java
@@ -2,20 +2,13 @@ package ai.verta.modeldb.config;
 
 import ai.verta.modeldb.ModelDBConstants;
 import ai.verta.modeldb.common.config.Config;
-import ai.verta.modeldb.common.config.DatabaseConfig;
 import ai.verta.modeldb.common.config.InvalidConfigException;
-import ai.verta.modeldb.common.config.RdbConfig;
 import ai.verta.modeldb.common.config.ServiceUserConfig;
 import ai.verta.modeldb.common.exceptions.InternalErrorException;
-import ai.verta.modeldb.common.futures.FutureGrpc;
 import ai.verta.modeldb.common.futures.FutureJdbi;
-import com.zaxxer.hikari.HikariDataSource;
-import com.zaxxer.hikari.metrics.prometheus.PrometheusMetricsTrackerFactory;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executor;
-import org.jdbi.v3.core.Jdbi;
 
 public class TestConfig extends Config {
   private static TestConfig config = null;
@@ -80,25 +73,8 @@ public class TestConfig extends Config {
     if (this.jdbi == null) {
       // Initialize HikariCP and jdbi
       final var databaseConfig = config.database;
-      this.jdbi = initializeJdbi(databaseConfig);
+      this.jdbi = initializeJdbi(databaseConfig, "modeldb_test");
     }
     return this.jdbi;
-  }
-
-  private FutureJdbi initializeJdbi(DatabaseConfig databaseConfig) {
-    final var hikariDataSource = new HikariDataSource();
-    final var dbUrl = RdbConfig.buildDatabaseConnectionString(databaseConfig.RdbConfiguration);
-    hikariDataSource.setJdbcUrl(dbUrl);
-    hikariDataSource.setUsername(databaseConfig.RdbConfiguration.RdbUsername);
-    hikariDataSource.setPassword(databaseConfig.RdbConfiguration.RdbPassword);
-    hikariDataSource.setMinimumIdle(Integer.parseInt(databaseConfig.minConnectionPoolSize));
-    hikariDataSource.setMaximumPoolSize(Integer.parseInt(databaseConfig.maxConnectionPoolSize));
-    hikariDataSource.setRegisterMbeans(true);
-    hikariDataSource.setMetricsTrackerFactory(new PrometheusMetricsTrackerFactory());
-    hikariDataSource.setPoolName("modeldb_test");
-
-    final Jdbi jdbi = Jdbi.create(hikariDataSource);
-    final Executor dbExecutor = FutureGrpc.initializeExecutor(databaseConfig.threadCount);
-    return new FutureJdbi(jdbi, dbExecutor);
   }
 }
